Create css-keyframes-spinner
Had to run it through SVGOMG & remove the xlink prefix (and namespace declaration) to get it under 1024 bytes. Still renders fine in modern browsers, of course!

PS, it wasn't clear, but if it should actually be under 1000 bytes, I can re-remove the line breaks. But since we expect people to read the code, I figured they were helpful!